requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aniso8601==8.1.0
attrs==20.3.0
click==7.1.2
Flask==1.1.2
Flask-HTTPAuth==4.4.0
Flask-JWT==0.3.2
Flask-HTTPAuth==4.4.0
flask-marshmallow==0.14.0
flask-restplus==0.13.0
flask-sqlalchemy==2.4.4
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
flask==1.1.2
importlib-metadata==3.1.1; python_version < '3.8'
itsdangerous==1.1.0
jinja2==2.11.2
jsonschema==3.2.0
markupsafe==1.1.1
marshmallow-sqlalchemy==0.24.1
marshmallow==3.9.1
pyrsistent==0.17.3
PyYAML==5.4.1
pytz==2020.4
six==1.15.0
sqlalchemy==1.3.20
werkzeug==0.16.1
zipp==3.4.0
